Feat - Image adding on vela done

+++ inventory.manager/src/main/java/com/velas/vivene/inventory/manager/repository/ImagemRepository.java
@@ -1,20 +1,22 @@
 package com.velas.vivene.inventory.manager.repository;
 
-import com.velas.vivene.inventory.manager.entity.Imagem;
-import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface ImagemRepository extends JpaRepository<Imagem, Long> {
+import com.velas.vivene.inventory.manager.entity.Imagem;
+
+import jakarta.transaction.Transactional;
+
+public interface ImagemRepository extends JpaRepository<Imagem, Integer> {
 
     @Query("update Imagem i set i.referencia = ?2 where i.id = ?1")
     @Modifying
     @Transactional
-    void updateReferenciaArquivoFoto(Long id, String referenciaArquivoFoto);
+    void updateReferenciaArquivoFoto(Integer id, String referenciaArquivoFoto);
 
 
     @Query("select i.referencia from Imagem i where i.id = ?1")
-    String findReferenciaArquivoFotoById(Long id);
+    String findReferenciaArquivoFotoById(Integer id);
 
 }
